refactor(backend): reorganização de pacotes, centralização de exceptions

+++ backend/src/apps/cursos/cursos_repository.py
@@ -1,14 +1,18 @@
+# Dependency
 from typing import Optional
-
-from sqlalchemy.orm import selectinload
-from sqlalchemy import select
-
-from src.err.exceptios import EntityNotFoundException
 from sqlalchemy import create_engine
-from src.database.base import Base
 from sqlalchemy.orm import sessionmaker
 
-from src.apps.cursos.cursos_model import CursosModel
+# Packages
+from . import CursosModel
+from src.database.base import Base
+
+# Exceptions
+from src.exceptions.process_error import (
+    UniqueConstraintViolationException,
+    EntityNotFoundException,
+)
+from sqlalchemy.exc import IntegrityError
 
 
 class CursosRepository:
@@ -22,9 +26,19 @@ class CursosRepository:
         self.session = self.Session()
 
     def save(self, entity_model: CursosModel) -> CursosModel:
-        self.session.add(entity_model)
-        self.session.commit()
-        return entity_model
+        try:
+            self.session.add(entity_model)
+            self.session.commit()
+            return entity_model
+
+        except IntegrityError as err:
+            self.session.rollback()
+            raise UniqueConstraintViolationException(err)
+
+        except Exception as err:
+            self.session.rollback()
+            print(type(err))
+            print(err)
 
     def find_all(self) -> list[CursosModel]:
         query = self.session.query(CursosModel)
@@ -32,14 +46,26 @@ class CursosRepository:
 
         return result
 
-    def find(self, _id: int) -> CursosModel | None:
-        result = self.session.query(CursosModel).filter_by(id=_id).first()
+    def find(self, _id: int = None, _externalId: str = None) -> CursosModel | None:
+
+        query = self.session.query(CursosModel)
+        result: Optional[CursosModel] = None
+
+        if _externalId is None:
+            result = query.filter_by(id=_id).first()
+        else:
+            result = query.filter(CursosModel.externalId == _externalId).first()
+
         if result is None:
-            raise EntityNotFoundException()
+            raise EntityNotFoundException("Curso não encontrado!")
+
         return result
 
     def edit(self, _id: int, entity_model: CursosModel) -> CursosModel | None:
         newEntity = self.session.query(CursosModel).filter_by(id=_id).first()
+
+        if newEntity is None:
+            raise EntityNotFoundException("Curso não encontrado!")
 
         newEntity.name = entity_model.name
         newEntity.externalId = entity_model.externalId
@@ -54,6 +80,10 @@ class CursosRepository:
 
     def remove(self, _id: int) -> CursosModel | None:
         resultEntity = self.session.query(CursosModel).filter_by(id=_id).first()
+
+        if resultEntity is None:
+            raise EntityNotFoundException("Curso não encontrado!")
+
         self.session.delete(resultEntity)
         self.session.commit()
         return resultEntity
